test: split test helpers
