+++ src/auth/dto/register-user.dto.spec.ts
@@ -12,7 +12,7 @@ import { RegisterUser } from '@/auth/dto/register-user.dto';
 import { User } from '@/auth/entities/user.entity';
 import { registerFixture } from '@/auth/fixtures/auth.fixture';
 import { IsAlreadyRegisterConstraint } from '@/auth/validators/is-already-register.validator';
-import { PASSWORD_HASHES } from '@/common/test-helpers';
+import { PASSWORD_HASHES } from '@/common/fixtures/password-hashes';
 
 describe('Register user validations', () => {
   beforeAll(async () => {
