+++ src/auth/factories/user.factory.ts
@@ -2,7 +2,7 @@ import normalizeEmail from 'normalize-email';
 import { setSeederFactory } from 'typeorm-extension';
 
 import { User } from '@/auth/entities/user.entity';
-import { credentials } from '@/common/test-helpers';
+import { credentials } from '@/common/fixtures/credentials';
 
 export const userFactory = setSeederFactory(User, (faker) =>
   User.fromPartial({
