+++ src/auth/fixtures/user.fixture.ts
@@ -2,9 +2,9 @@ import { fluse } from 'fluse';
 import normalizeEmail from 'normalize-email';
 
 import { User } from '@/auth/entities/user.entity';
+import { credentials } from '@/common/fixtures/credentials';
 import { fakerPlugin } from '@/common/fluse-plugin-faker';
 import { typeORMPlugin } from '@/common/fluse-plugin-typeorm';
-import { credentials } from '@/common/test-helpers';
 
 const { fixture } = fluse({
   plugins: {
