+++ src/common/database.ts
@@ -1,0 +1,18 @@
+import { randomUUID } from 'node:crypto';
+import { DataType, newDb } from 'pg-mem';
+
+export const database = newDb({
+  autoCreateForeignKeyIndices: true,
+});
+
+database.public.registerFunction({
+  name: 'current_database',
+  implementation: () => 'test',
+});
+
+database.public.registerFunction({
+  name: 'uuid_generate_v4',
+  returns: DataType.uuid,
+  implementation: () => randomUUID(),
+  impure: true,
+});
