+++ src/common/fixtures/password-hashes.ts
@@ -1,0 +1,12 @@
+export const PASSWORD_HASHES = `
+01B0E792722BA73BB221BB447CAD2E95B5A:21
+01D9ACFB1F9BDF8593B2A35F043745AD9FB:1
+040911493F28B1CD9EFDBECB96CE2A94324:6
+074B0C6691134C16E0603ABE4C776EEB75C:42
+076FBD9B37A06158F72C7B0B1118A9535BD:26
+098FA0B1B16B20E37E8B3E829C692C50B2A:1205
+1631F412B2136652FD8A11060E324B11F9B:3
+164C1011FEF7893941E58C62666827BD879:15
+1E4C9B93F3F0682250B6CF8331B7EE68FD8:9545824
+A8B6421D0C680037FA68FE7FD266C4FE2F5:1805
+`;
