+++ src/common/test-matchers.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line security/detect-unsafe-regex
+export const uuidRegex = /[\da-f]{8}(?:-[\da-f]{4}){3}-[\da-f]{12}/;
+
+export const isoDateRegex = /\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z/;
